Removing uses of long long type in the C++ backend (only remaining instance is now Amir's hash function which should be changed to a 32-bit implementation, e.g. in MurmurHash3).
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@2616 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 40b05759508fb42b0a140152cc03a75b6051dffb

+++ lib/dbt_c++/standard_adaptors.hpp
@@ -14,7 +14,7 @@ namespace dbtoaster {
   namespace adaptors {
 
     using namespace std;
-	using namespace boost;
+    using namespace boost;
     using namespace dbtoaster;
     using namespace dbtoaster::streams;
 
@@ -75,7 +75,7 @@ namespace dbtoaster {
       // Struct to represent messages coming off a socket/historical file
       struct order_book_message {
           double t;
-          long long id;
+          long id;
           string action;
           double volume;
           double price;
@@ -85,8 +85,8 @@ namespace dbtoaster {
       // order id.
       struct order_book_tuple {
           double t;
-          long long id;
-          long long broker_id;
+          long id;
+          long broker_id;
           double volume;
           double price;
           order_book_tuple() {}
